Changing pages to ssr part 4: layout [nodeploy]
